requirements-dev.txt
opencv-python==3.4.8.29
tensorflow==2.8
flask==2.2
quilt
boto3
python-chess
papermill
imbalanced-learn
matplotlib
protobuf==3.20.1
wandb
aquariumlearning
setuptools
cairosvg
pipwin
#pipwin install cairocffi
ipython
ipykernel
